refactor: split sample place generator responsibilities
